Remove code with side-affects
These few lines caused side-affects on other pages.
This seems to have originally been as a quick and dirty fix for the threads and/or forums but it affects way more pages than that.

Next commit should fix many pages and their consistency and thus removes the need for these lines.